requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
gunicorn==21.2.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
PyJWT==2.8.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
zipp==3.16.2
